# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Application API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/employees/{company_id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get all Employees",
                "description": "Get all employees of one company by company ID , this route worked  for admin and company roles.",
                "operationId": "getEmployees",
                "parameters": [
                    {
                        "name": "company_id",
                        "in": "path",
                        "description": "Company id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/employee/{employee_id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get employee by ID",
                "description": "Get one employee with ID, this route worked for admin and company roles.",
                "operationId": "getemployee",
                "parameters": [
                    {
                        "name": "employee_id",
                        "in": "path",
                        "description": "Employee id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/employee/create": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Create employee",
                "description": "Create employee, this route worked only for admin and company roles.",
                "operationId": "addEmployee",
                "requestBody": {
                    "description": "Create employee",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "passport_serial",
                                    "first_name",
                                    "address",
                                    "father_name",
                                    "last_name",
                                    "position",
                                    "phone",
                                    "company_id"
                                ],
                                "properties": {
                                    "passport_serial": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "AA777999"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Oliver"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Jake"
                                    },
                                    "father_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Mike"
                                    },
                                    "position": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Manager"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Uzbekistan Tashkent region, Parkent town"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3333333333"
                                    },
                                    "company_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Company"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/employee/update/{employee_id}": {
            "put": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get employee by ID",
                "description": "Update employee with ID, this route worked for admin and company roles.",
                "operationId": "updateEmployee",
                "parameters": [
                    {
                        "name": "employee_id",
                        "in": "path",
                        "description": "Employee id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Create employee",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "passport_serial",
                                    "first_name",
                                    "address",
                                    "father_name",
                                    "last_name",
                                    "position",
                                    "phone",
                                    "company_id"
                                ],
                                "properties": {
                                    "passport_serial": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "AA777999"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Oliver"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Jake"
                                    },
                                    "father_name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Mike"
                                    },
                                    "position": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Manager"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Uzbekistan Tashkent region, Parkent town"
                                    },
                                    "phone": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3333333333"
                                    },
                                    "company_id": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/employee/delete/{employee_id}": {
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Delete employee by ID",
                "description": "Delete employee with ID, this route worked only for admin and company roles.",
                "operationId": "deleteEmployee",
                "parameters": [
                    {
                        "name": "employee_id",
                        "in": "path",
                        "description": "Employee id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Company": {
                "title": "Company",
                "description": "Company model, companies can be changed or added only by admins.",
                "xml": {
                    "name": "Company"
                }
            },
            "Employee": {
                "title": "Employee",
                "description": "Employee model, employees can be changed or added by admins and users that have company role.",
                "xml": {
                    "name": "Employee"
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}